:bug: fixed depth image accuracy based on cdowen issue

+++ mapper.py
@@ -54,8 +54,8 @@ def depth_2_color_space(kinect, depth_space_point, depth_frame_data, show=False,
     depthXs = np.clip(depthXYs[:, :, 0], 0, kinect.depth_frame_desc.Width - 1)
     depthYs = np.clip(depthXYs[:, :, 1], 0, kinect.depth_frame_desc.Height - 1)
     depth_frame = kinect.get_last_depth_frame()
-    depth_img = depth_frame.reshape((kinect.depth_frame_desc.Height, kinect.depth_frame_desc.Width, 1)).astype(np.uint8)
-    align_depth_img = np.zeros((1080, 1920, 4), dtype=np.uint8)
+    depth_img = depth_frame.reshape((kinect.depth_frame_desc.Height, kinect.depth_frame_desc.Width, 1)).astype(np.uint16)
+    align_depth_img = np.zeros((1080, 1920, 4), dtype=np.uint16)
     align_depth_img[:, :] = depth_img[depthYs, depthXs, :]
     if show:
         cv2.imshow('Aligned Image', cv2.resize(cv2.flip(align_depth_img, 1), (int(1920 / 2.0), int(1080 / 2.0))))
